heart_stretch

+++ projects/heartStretch/26012023_followup_screen/run_scripts/run_multiple_simulation.py
@@ -67,7 +67,7 @@ def run_simulation(path_name,p_notch_idx,seed_idx,set_idx):
                         grn_params=scan_dict["grn_params"],
                         run_options=scan_dict["run_options"],
                         save_options=scan_dict["save_options"])
-    sim.t.mesh.x = x.astype(np.int32)
+    sim.t.mesh.x = x.astype(np.float32)
     sim.grn.is_notch = is_notch.astype(bool)
     sim.simulate(progress_bar=False)
